tests/submission/W3C/domhtml/HTMLIFrameElement04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IFRAME</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-iframe-longdesc" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<iframe height="50" longdesc="http://www.w3.org/" src="/support/html/sample.html" width="60">IFRAME1</iframe>


<script>
var domtest = async_test("The longDesc IDL attribute of the frame element must reflect the element's longdesc content attribute, which for the purposes of reflection is defined as containing a URL.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vlongdesc;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("iframe");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vlongdesc = testNode.longDesc;

      assert_equals(vlongdesc,"http://www.w3.org/");
       
    domtest.done();
}));

</script>
<p>The longDesc IDL attribute of the frame element must reflect the element's longdesc content attribute, which for the purposes of reflection is defined as containing a URL.</p>
<div id="log"></div>
</body></html>
